Add new function to create an instance of MimiRICE2010
This commit adds a function to create a Mimi RICE2010 model and also adds a file to calculate welfare without Negishi weights. The function has a true/false switch for whether we are using Negishi weights, and will swap in the new welfare component if they are turned off. This was necessary because the pure rate of time preference is fixed at 0.015 in the Mimi package of RICE2010, so not possible to modify that paramter in sensitivity tests.